fix: write story in new format
